refactor: introduced the concept of `BuildSteps`
Split the larger `Java|Groovy|KotlinBuilder` classes up into smaller
`BuildStep` classes, each handling a very well-defined subset of the
build process.
