+++ src/main/java/dev/jbang/source/Builder.java
@@ -2,6 +2,6 @@ package dev.jbang.source;
 
 import java.io.IOException;
 
-public interface Builder {
-	Project build() throws IOException;
+public interface Builder<T> {
+	T build() throws IOException;
 }
